refactor(gui): split App.tsx per-tab JSX into views/*View.tsx
App.tsx had grown to 1480 lines holding the full JSX for 7 tool tabs
(discover, scan, inspect, sweep, dns, interfaces, pcap) plus the
dashboard and settings tabs. The renderToolbar() and renderContent()
switch statements between them were 380 + 300 lines of inline
JSX with significant duplication of the form/input/button pattern.

Split each tool tab into its own file under views/ with two named
exports: XxxToolbar (the form + run button) and XxxResults (the
result renderer). DashboardView and SettingsView were already in
this shape. The new files:

  views/DiscoverView.tsx    78 lines  (subnet input + host table)
  views/ScanView.tsx       115 lines  (host+ports + open-ports table)
  views/InspectView.tsx     94 lines  (host+ports + JSON dump)
  views/SweepView.tsx      111 lines  (subnet+ports + host cards)
  views/DnsView.tsx        117 lines  (host+type + records table)
  views/InterfacesView.tsx 142 lines  (interfaces table + ARP table)
  views/PcapView.tsx       190 lines  (pcap form + summary card)

Each view file is self-contained: it imports the relevant netscli
service function, defines its prop interface, and exports the two
JSX components. Per-tab validation is now inline in the toolbar
component instead of using the shared validateRequired helper.

App.tsx (now 931 lines) keeps:
- Cross-tab state: tabOutputs, tabErrors, tabBusy, tabViewMode,
  scanHistory, tabActiveHistoryId
- Lifted form state: subnet, host, ports, dnsType, pcapInterface,
  pcapDuration (still cross-tab so e.g. typing a host on Scan
  pre-fills it on Inspect)
- The runOperation/cancelOperation lifecycle (per-tab opSeq
  tracking, mountedRef, currentOpId)
- TitleBar menus, network stats polling, exports, history list

renderToolbar() shrinks from ~380 lines of inline JSX to a 70-line
switch dispatching to <XxxToolbar /> components. renderContent()
shrinks from ~300 lines to a 30-line switch dispatching to
<XxxResults /> components on output.kind.

The pcapGuidance() helper moved to PcapView.tsx so the pcap-specific
error-hint logic lives next to the pcap view that uses it.

Verified locally:
- npm run build (tsc + vite) clean — 41 modules transformed
  (was 34); bundle size 239 KB (was 237 KB), gzip 71 KB (was 70 KB)
- cargo check -p netscli-gui clean (Tauri shell unaffected)
